--- apps/trustlab/src/components/DonorOverview/OurDonors.snap.js
@@ -1,3 +1,0 @@
-// Jest Snapshot v1, https://goo.gl/fbAQLP
-
-exports[`<DonorOverview /> renders unchanged 1`] = `<div />`;
