--- apps/trustlab/src/components/DonorOverview/index.js
@@ -1,3 +1,0 @@
-import DonorOverview from "./DonorOverview";
-
-export default DonorOverview;
