+++ apps/trustlab/src/components/DonorOverviewList/DonorOverviewList.snap.js
@@ -1,0 +1,3 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`<DonorOverviewList /> renders unchanged 1`] = `<div />`;
