+++ apps/trustlab/src/components/DonorOverviewList/index.js
@@ -1,0 +1,3 @@
+import DonorOverviewList from "./DonorOverviewList";
+
+export default DonorOverviewList;
