--- apps/trustlab/src/components/PartnerOverview/PartnerOverview.snap.js
@@ -1,3 +1,0 @@
-// Jest Snapshot v1, https://goo.gl/fbAQLP
-
-exports[`<PartnerOverview /> renders unchanged 1`] = `<div />`;
