--- apps/trustlab/src/components/PartnerOverview/index.js
@@ -1,3 +1,0 @@
-import PartnerOverview from "./PartnerOverview";
-
-export default PartnerOverview;
