+++ apps/trustlab/src/components/PartnerOverviewList/PartnerOverviewList.snap.js
@@ -1,0 +1,3 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`<PartnerOverviewList /> renders unchanged 1`] = `<div />`;
