+++ apps/trustlab/src/components/PartnerOverviewList/index.js
@@ -1,0 +1,3 @@
+import PartnerOverviewList from "./PartnerOverviewList";
+
+export default PartnerOverviewList;
